+++ Meshconvert/Meshconvert_Desktop_2019_Win10.vcxproj
@@ -420,6 +420,7 @@
     <ClCompile Include="Mesh.cpp" />
     <ClCompile Include="MeshOBJ.cpp" />
     <ClInclude Include="..\Utilities\WaveFrontReader.h" />
+    <ClInclude Include="CmdLineHelpers.h" />
     <CLInclude Include="Mesh.h" />
     <ClInclude Include="SDKMesh.h" />
   </ItemGroup>
